Added --profile, --watch, --breakpoint, --debug as global options (#917)
